feat: add proper completion test with a completion pool
This should help preventing cahcing and invalidating previous results

Previous results showed an average response time of ~500ms for 100 users
making approx 15 req/s which raised some suspects.
After this new tests, more realistic insights on both the proxy and the
LSP are coming out, due to the new results of an average of 2500ms for a
response under such workload.